+++ flocker/volume/filesystems/zfs.py
@@ -514,6 +514,22 @@ class StoragePool(Service):
         d.addCallback(lambda _: filesystem)
         return d
 
+    def set_maximum_size(self, volume):
+        filesystem = self.get(volume)
+        properties = []
+        if volume.size.maximum_size is not None:
+            properties.extend([
+                u"refquota={0}".format(
+                    volume.size.maximum_size).encode("ascii")
+            ])
+        else:
+            properties.extend([u"refquota=none"])
+        d = zfs_command(self._reactor,
+                        [b"set"] + properties + [filesystem.name])
+        d.addErrback(self._check_for_out_of_space)
+        d.addCallback(lambda _: filesystem)
+        return d
+
     def clone_to(self, parent, volume):
         parent_filesystem = self.get(parent)
         new_filesystem = self.get(volume)
@@ -648,11 +664,9 @@ def _list_filesystems(reactor, pool):
             name, mountpoint, refquota = line.split(b'\t')
             name = name[len(pool) + 1:]
             if name:
-                if refquota == b"none":
+                refquota = int(refquota.decode("ascii"))
+                if refquota == 0:
                     refquota = None
-                else:
-                    refquota = int(refquota.decode("ascii"))
-                    refquota = None if refquota == 0 else refquota
                 yield _DatasetInfo(
                     dataset=name, mountpoint=mountpoint, refquota=refquota)